% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\docType{package}
\name{nfer}
\alias{nfer}
\title{nfer: A package for inferring interval abstractions of event traces.}
\description{
The nfer package provides an R interface to the nfer language for
abstracting event traces into a hierarchy of intervals with data.
}
\section{nfer functions}{

load, learn, read, apply
}

